simple solution

+++ RaceCars/src/com/beck/racecars/RedCar.java
@@ -2,7 +2,7 @@ package com.beck.racecars;
 
 public class RedCar extends BasicCar{
 	
-	public RedCar (int weight, int power)
+	public RedCar (float weight, float power)
 	{
 	super(weight, power);
 	}
